services can only takeaway or delivery

+++ routes/restaurants.js
@@ -1,150 +1,102 @@
-const express = require("express");
-const router = express.Router();
-const authenticateToken = require("../middlewares/authenticateToken");
-const checkBody = require("../utils/checkBody");
-const Restaurant = require("../models/restaurant");
-const Food = require("../models/food");
-const Option = require("../models/option");
+const express = require("express")
+const router = express.Router()
+const authenticateToken = require("../middlewares/authenticateToken")
+const checkBody = require("../utils/checkBody")
+const Restaurant = require("../models/restaurant")
+const Food = require("../models/food")
+const Option = require("../models/option")
+const checkEnvironment = require("../middlewares/checkEnvironment")
 
-const catchAsyncErrors = require("../utils/catchAsyncErrors");
-const AppError = require("../utils/AppError");
+const catchAsyncErrors = require("../utils/catchAsyncErrors")
+const AppError = require("../utils/AppError")
 
 //utils
 const generateUniqueValue = async (name, city) => {
   //check if restaurant name exists as uniqueValue in the db, if exists try name-city, if exists try name-city-index
-  let uniqueValue = name.toLowerCase();
-  let i = 1;
+  let uniqueValue = name.toLowerCase()
+  let i = 1
 
   while (true) {
-    const existingRestaurant = await Restaurant.findOne({ uniqueValue });
+    const existingRestaurant = await Restaurant.findOne({ uniqueValue })
 
     if (!existingRestaurant) {
-      break;
+      break
     }
 
     if (i === 1) {
-      uniqueValue = `${name}-${city}`.toLowerCase();
+      uniqueValue = `${name}-${city}`.toLowerCase()
     } else {
-      uniqueValue = `${name}-${city}-${i}`.toLowerCase();
+      uniqueValue = `${name}-${city}-${i}`.toLowerCase()
     }
 
-    i++;
+    i++
   }
 
-  return uniqueValue;
-};
+  return uniqueValue
+}
 
-createMenu = async (menu, restaurantUniqueValue) => {
-  const populatedMenu = await Promise.all(
-    menu.map(async (foodCategory, i) => {
-      const foods = [];
-      for (const food of foodCategory.foods) {
-        const optionGroups = [];
-        for (const optionGroup of food.optionGroups) {
-          const optionIds = [];
-          for (const option of optionGroup.options) {
-            const optionFound = await Option.findOne({
-              value: option.value,
-              isSupplement: option.isSupplement,
-              price: option.price,
-              restaurantUniqueValue: restaurantUniqueValue,
-            });
-            if (optionFound) {
-              optionIds.push(optionFound._id);
-            } else {
-              const newOption = await Option.create({
-                ...option,
-                restaurantUniqueValue: restaurantUniqueValue,
-              });
-              optionIds.push(newOption._id);
-            }
+const createMenu = async (menu, restaurantUniqueValue) => {
+  const populatedMenu = []
+  for (let i = 0; i < menu.length; i++) {
+    const foodCategory = menu[i]
+    const foods = []
+    for (let j = 0; j < foodCategory.foods.length; j++) {
+      const food = foodCategory.foods[j]
+      const optionGroups = []
+      for (let k = 0; k < food.optionGroups.length; k++) {
+        const optionGroup = food.optionGroups[k]
+        const optionIds = []
+        for (let m = 0; m < optionGroup.options.length; m++) {
+          const option = optionGroup.options[m]
+          //avoid no finding option because IsSupplement is null
+          if (!option.isSupplement) {
+            option.isSupplement = false
           }
-          optionGroup.options = optionIds;
-          optionGroups.push(optionGroup);
-        }
-
-        const newFood = await Food.create({
-          ...food,
-          categoryNumber: i,
-          categoryTitle: foodCategory.title,
-          optionGroups: optionGroups,
-          restaurantUniqueValue: restaurantUniqueValue,
-        });
-        foods.push(newFood._id);
-      }
-      return {
-        ...foodCategory,
-        foods,
-      };
-    })
-  );
-  return populatedMenu;
-};
-
-updateMenu = async (menu, restaurantUniqueValue) => {
-  const populatedMenu = await Promise.all(
-    menu.map(async (foodCategory, i) => {
-      const foods = [];
-      for (const food of foodCategory.foods) {
-        const optionGroups = [];
-        for (const optionGroup of food.optionGroups) {
-          const optionIds = [];
-          for (const option of optionGroup.options) {
-            const optionFound = await Option.findOne({
-              _id: option._id,
-            });
-            if (optionFound) {
-              const { _id, ...optionData } = option;
-              // Merge the properties of optionData into optionFound
-              Object.assign(optionFound, optionData);
-              optionFound.save();
-              optionIds.push(optionFound.ids);
-            } else {
-              const newOption = await Option.create({
-                ...option,
-                restaurantUniqueValue: restaurantUniqueValue,
-              });
-              optionIds.push(newOption._id);
-            }
-          }
-          optionGroup.options = optionIds;
-          optionGroups.push(optionGroup);
-        }
-
-        const foodFound = await Food.findOne({ _id: food._id });
-        if (foodFound) {
-          const { _id, ...foodData } = food;
-          // Merge the properties of optionData into optionFound
-          Object.assign(optionFound, foodData);
-          (foodFound.categoryTitle = foodCategory.title),
-            (foodFound.categoryNumber = i);
-          foodFound.save();
-          foods.push(foodFound.ids);
-        } else {
-          const newFood = await Food.create({
-            ...food,
-            categoryNumber: i,
-            categoryTitle: foodCategory.title,
-            optionGroups: optionGroups,
+          const optionFound = await Option.findOne({
+            value: option.value,
+            isSupplement: option.isSupplement,
+            price: option.price,
             restaurantUniqueValue: restaurantUniqueValue,
-          });
-          foods.push(newFood._id);
+          })
+          let optionId
+          if (optionFound) {
+            console.log("yes found !")
+            optionId = optionFound._id
+          } else {
+            const newOption = await Option.create({
+              ...option,
+              restaurantUniqueValue: restaurantUniqueValue,
+            })
+            optionId = newOption._id
+          }
+          optionIds.push(optionId)
         }
+        optionGroup.options = optionIds
+        optionGroups.push(optionGroup)
       }
-      return {
-        ...foodCategory,
-        foods,
-      };
+
+      const newFood = await Food.create({
+        ...food,
+        categoryNumber: i,
+        categoryTitle: foodCategory.title,
+        optionGroups: optionGroups,
+        restaurantUniqueValue: restaurantUniqueValue,
+      })
+      foods.push(newFood._id)
+    }
+    populatedMenu.push({
+      ...foodCategory,
+      foods,
     })
-  );
-  return populatedMenu;
-};
+  }
+  return populatedMenu
+}
 
 //routes
 // create a new restaurant Document, this requet is IP limited
-router.post("/addRestaurant", async (req, res, next) => {
-  const allowedIPs = ["::ffff:127.0.0.1"];
-  const clientIP = req.ip;
+router.post("/addRestaurant", checkEnvironment, async (req, res, next) => {
+  const allowedIPs = ["::ffff:127.0.0.1"]
+  const clientIP = req.ip
   if (
     !checkBody(req.body, [
       "name",
@@ -157,22 +109,22 @@ router.post("/addRestaurant", async (req, res, next) => {
       "menu",
     ])
   ) {
-    throw new AppError("Body is incorrect", 400, "BadRequestError");
+    throw new AppError("Body is incorrect", 400, "BadRequestError")
   }
   if (!allowedIPs.includes(clientIP)) {
     throw new AppError(
       "This IP address is not allowed for this route",
       400,
       "ForbiddenError"
-    );
+    )
   }
 
   const uniqueValue = await generateUniqueValue(
     req.body.name,
     req.body.address.city
-  );
+  )
 
-  const menuWithFoodIds = await createMenu(req.body.menu);
+  const menuWithFoodIds = await createMenu(req.body.menu)
 
   const newRestaurant = await Restaurant.create({
     name: req.body.name,
@@ -184,10 +136,10 @@ router.post("/addRestaurant", async (req, res, next) => {
     publicSettings: req.body.publicSettings,
     privateSettings: req.body.privateSettings,
     menu: menuWithFoodIds,
-  });
+  })
 
-  return res.json(newRestaurant);
-});
+  return res.json(newRestaurant)
+})
 
 //public restaurant data
 router.get(
@@ -196,7 +148,7 @@ router.get(
     const restaurant = await Restaurant.findOne({
       uniqueValue: req.params.uniqueValue.toLowerCase(),
     })
-      .select("-privateSettings -_id")
+      .select("-privateSettings")
       .populate({
         path: "menu.foods",
         match: { display: true },
@@ -206,17 +158,17 @@ router.get(
             path: "options",
           },
         },
-      });
+      })
 
     if (!restaurant) {
-      throw new AppError("Restaurant Not Found", 404, "NotFoundError");
+      throw new AppError("Restaurant Not Found", 404, "NotFoundError")
     } else {
       //only keep element withtiongroup display = true for public data
-      restaurant.menu = restaurant.menu;
-      return res.json(restaurant);
+      restaurant.menu = restaurant.menu
+      return res.json(restaurant)
     }
   })
-);
+)
 
 //public restaurant data
 router.get(
@@ -225,7 +177,7 @@ router.get(
   catchAsyncErrors(async (req, res, next) => {
     const restaurant = await Restaurant.findOne({
       uniqueValue: req.params.uniqueValue.toLowerCase(),
-    });
+    })
     //
     if (restaurant) {
       //check if restaurantUniqueValue inside the jwt token is the same as restaurant.uniqueValue
@@ -234,14 +186,14 @@ router.get(
           "data access is Forbidden for this user",
           403,
           "ForbiddenError"
-        );
+        )
       }
-      return res.json(restaurant);
+      return res.json(restaurant)
     } else {
-      throw new AppError("Restaurant Not Found", 404, "NotFoundError");
+      throw new AppError("Restaurant Not Found", 404, "NotFoundError")
     }
   })
-);
+)
 
 router.post(
   "/admin/updateRestaurantSettings",
@@ -249,7 +201,7 @@ router.post(
   catchAsyncErrors(async (req, res, next) => {
     const restaurant = await Restaurant.findOne({
       uniqueValue: req.body.uniqueValue,
-    });
+    })
     if (restaurant) {
       //check if restaurantUniqueValue inside the jwt token is the same as restaurant.uniqueValue
       if (req.user.restaurantUniqueValue !== restaurant.uniqueValue) {
@@ -257,75 +209,64 @@ router.post(
           "Update restaurant settings is Forbidden for this user",
           403,
           "ForbiddenError"
-        );
+        )
       }
       //avoid deliveryPostCodes containing empty string
       req.body.publicSettings.deliveryPostCodes =
         req.body.publicSettings.deliveryPostCodes.filter(
           (postCode) => postCode !== ""
-        );
+        )
 
-      restaurant.name = req.body.name;
-      restaurant.mail = req.body.mail;
-      restaurant.address = req.body.address;
-      restaurant.phoneNumber = req.body.phoneNumber;
-      restaurant.website = req.body.website;
-      restaurant.publicSettings = req.body.publicSettings;
-      restaurant.privateSettings = req.body.privateSettings;
+      restaurant.name = req.body.name
+      restaurant.mail = req.body.mail
+      restaurant.address = req.body.address
+      restaurant.phoneNumber = req.body.phoneNumber
+      restaurant.website = req.body.website
+      restaurant.publicSettings = req.body.publicSettings
+      restaurant.privateSettings = req.body.privateSettings
       restaurant.save().then((savedRestaurant) => {
         // Vérifiez si l'enregistrement a été sauvegardé avec succès
         if (savedRestaurant === restaurant) {
-          return res.json(savedRestaurant);
+          return res.json(savedRestaurant)
         } else {
-          throw new AppError("Failed to save restaurant settings");
+          throw new AppError("Failed to save restaurant settings")
         }
-      });
+      })
     } else {
-      throw new AppError("Restaurant Not Found", 404, "NotFoundError");
+      throw new AppError("Restaurant Not Found", 404, "NotFoundError")
     }
   })
-);
+)
 
-//delete restaurant and all document associated with
-
-/*router.delete(
-  "/:restaurantId",
-  catchAsyncErrors(async (req, res, next) => {
-    removeFoodRestaurant(req.params.restaurantId,true);
-    res.json("Le restaurant a été supprimé avec succès");
-  })
-);
-*/
-
-/*router.delete(
-  "menu/:restaurantId",
-  catchAsyncErrors(async (req, res, next) => {
-    removeFoodRestaurant(req.params.restaurantId,false);
-    res.json("Le menu a été supprimé avec succès");
-  })
-);
-*/
-
-/*
 router.post(
   "/createMenu",
+  checkEnvironment,
   catchAsyncErrors(async (req, res, next) => {
+    const allowedIPs = ["::ffff:127.0.0.1"]
+    const clientIP = req.ip
+    if (!allowedIPs.includes(clientIP)) {
+      throw new AppError(
+        "This IP address is not allowed for this route",
+        400,
+        "ForbiddenError"
+      )
+    }
     const restaurantFound = await Restaurant.findOne({
       uniqueValue: req.body.uniqueValue,
-    });
+    })
     if (!restaurantFound) {
-      throw new AppError("Restaurant Not Foud", 404, "ErrorNotFound");
+      throw new AppError("Restaurant Not Foud", 404, "ErrorNotFound")
     }
     // delete the foods linked to Restaurant
-    await Food.deleteMany({ restaurantUniqueValue: req.body.uniqueValue });
-    await Option.deleteMany({ restaurantUniqueValue: req.body.uniqueValue });
-    const newMenu = await createMenu(req.body.menu, req.body.uniqueValue);
-    restaurantFound.menu = newMenu;
-    restaurantFound.save();
-    res.json("Le menu a été modifié avec succès");
+    await Food.deleteMany({ restaurantUniqueValue: req.body.uniqueValue })
+    await Option.deleteMany({ restaurantUniqueValue: req.body.uniqueValue })
+    const newMenu = await createMenu(req.body.menu, req.body.uniqueValue)
+    restaurantFound.menu = newMenu
+    restaurantFound.save()
+    res.json("Le menu a été modifié avec succès")
   })
-);
-*/
+)
+
 /*
 router.post(
   "/updateMenu",
@@ -354,4 +295,4 @@ router.post(
 );
 */
 
-module.exports = router;
+module.exports = router
